Move name generator and formula API to base Lua kernel
