+++ src/scripting/game_lua_kernel.cpp
@@ -74,7 +74,6 @@
 #include "scripting/lua_api.hpp"        // for luaW_toboolean, etc
 #include "scripting/lua_common.hpp"
 #include "scripting/lua_cpp_function.hpp"
-#include "scripting/lua_formula_bridge.hpp"
 #include "scripting/lua_gui2.hpp"	// for show_gamestate_inspector
 #include "scripting/lua_pathfind_cost_calculator.hpp"
 #include "scripting/lua_race.hpp"
@@ -104,8 +103,6 @@
 #include "units/ptr.hpp"                 // for unit_const_ptr, unit_ptr
 #include "units/types.hpp"    // for unit_type_data, unit_types, etc
 #include "util.hpp"                     // for lexical_cast
-#include "utils/markov_generator.hpp"
-#include "utils/context_free_grammar_generator.hpp"
 #include "variable.hpp"                 // for vconfig, etc
 #include "variable_info.hpp"
 #include "wml_exception.hpp"
@@ -4454,69 +4451,6 @@ int game_lua_kernel::intf_toggle_fog(lua_State *L, const bool clear)
 	return 0;
 }
 
-static int intf_name_generator(lua_State *L)
-{
-	std::string type = luaL_checkstring(L, 1);
-	name_generator* gen = nullptr;
-	if(type == "markov" || type == "markov_chain") {
-		std::vector<std::string> input;
-		if(lua_istable(L, 2)) {
-			input = lua_check<std::vector<std::string>>(L, 2);
-		} else {
-			input = utils::parenthetical_split(luaL_checkstring(L, 2));
-		}
-		int chain_sz = luaL_optinteger(L, 3, 2);
-		int max_len = luaL_optinteger(L, 4, 12);
-		gen = new(lua_newuserdata(L, sizeof(markov_generator)))
-			markov_generator(input, chain_sz, max_len);
-		// Ensure the pointer didn't change when cast
-		assert(static_cast<void*>(gen) == dynamic_cast<markov_generator*>(gen));
-	} else if(type == "context_free" || type == "cfg" || type == "CFG") {
-		void* buf = lua_newuserdata(L, sizeof(context_free_grammar_generator));
-		if(lua_istable(L, 2)) {
-			std::map<std::string, std::vector<std::string>> data;
-			for(lua_pushnil(L); lua_next(L, 2); lua_pop(L, 1)) {
-				if(!lua_isstring(L, -2)) {
-					lua_pushstring(L, "CFG generator: invalid nonterminal name (must be a string)");
-					return lua_error(L);
-				}
-				if(lua_isstring(L, -1)) {
-					data[lua_tostring(L,-2)] = utils::split(lua_tostring(L,-1),'|');
-				} else if(lua_istable(L, -1)) {
-					data[lua_tostring(L,-2)] = lua_check<std::vector<std::string>>(L, -1);
-				} else {
-					lua_pushstring(L, "CFG generator: invalid noterminal value (must be a string or list of strings)");
-					return lua_error(L);
-				}
-			}
-			if(!data.empty()) {
-				gen = new(buf) context_free_grammar_generator(data);
-			}
-		} else {
-			gen = new(buf) context_free_grammar_generator(luaL_checkstring(L, 2));
-		}
-		if(gen) {
-			assert(static_cast<void*>(gen) == dynamic_cast<context_free_grammar_generator*>(gen));
-		}
-	} else {
-		return luaL_argerror(L, 1, "should be either 'markov_chain' or 'context_free'");
-	}
-	static const char*const generic_err = "error initializing name generator";
-	if(!gen) {
-		lua_pushstring(L, generic_err);
-		return lua_error(L);
-	}
-	// We set the metatable now, even if the generator is invalid, so that it
-	// will be properly collected if it was invalid.
-	luaL_getmetatable(L, "name generator");
-	lua_setmetatable(L, -2);
-	if(!gen->is_valid()) {
-		lua_pushstring(L, generic_err);
-		return lua_error(L);
-	}
-	return 1;
-}
-
 // END CALLBACK IMPLEMENTATION
 
 game_board & game_lua_kernel::board() {
@@ -4586,20 +4520,17 @@ game_lua_kernel::game_lua_kernel(CVideo * video, game_state & gs, play_controlle
 		{ "add_known_unit",           &intf_add_known_unit           },
 		{ "add_modification",         &intf_add_modification         },
 		{ "advance_unit",             &intf_advance_unit             },
-		{ "compile_formula",          &lua_formula_bridge::intf_compile_formula},
 		{ "copy_unit",                &intf_copy_unit                },
 		{ "create_unit",              &intf_create_unit              },
 		{ "debug",                    &intf_debug                    },
 		{ "debug_ai",                 &intf_debug_ai                 },
 		{ "eval_conditional",         &intf_eval_conditional         },
-		{ "eval_formula",             &lua_formula_bridge::intf_eval_formula},
 		{ "get_era",                  &intf_get_era                  },
 		{ "get_image_size",           &intf_get_image_size           },
 		{ "get_time_stamp",           &intf_get_time_stamp           },
 		{ "get_traits",               &intf_get_traits               },
 		{ "get_viewing_side",         &intf_get_viewing_side         },
 		{ "modify_ai",                &intf_modify_ai                },
-		{ "name_generator",           &intf_name_generator           },
 		{ "set_music",                &intf_set_music                },
 		{ "transform_unit",           &intf_transform_unit           },
 		{ "unit_ability",             &intf_unit_ability             },
@@ -4794,9 +4725,6 @@ game_lua_kernel::game_lua_kernel(CVideo * video, game_state & gs, play_controlle
 	lua_pushstring(L, "unit variables");
 	lua_setfield(L, -2, "__metatable");
 	lua_rawset(L, LUA_REGISTRYINDEX);
-	
-	// Create formula bridge metatables
-	cmd_log_ << lua_formula_bridge::register_metatables(L);
 
 	// Create the vconfig metatable.
 	cmd_log_ << lua_common::register_vconfig_metatable(L);
