+++ src/scripting/lua_race.cpp
@@ -17,7 +17,6 @@
 #include "units/race.hpp"
 #include "scripting/lua_common.hpp"
 #include "units/types.hpp"
-#include "utils/name_generator.hpp"
 
 #include <string>
 #include <string.h>
@@ -92,20 +91,6 @@ static int impl_race_get(lua_State* L)
 	return 0;
 }
 
-static int impl_name_generator_call(lua_State *L)
-{
-	name_generator* gen = static_cast<name_generator*>(lua_touserdata(L, 1));
-	lua_pushstring(L, gen->generate().c_str());
-	return 1;
-}
-
-static int impl_name_generator_collect(lua_State *L)
-{
-	name_generator* gen = static_cast<name_generator*>(lua_touserdata(L, 1));
-	gen->~name_generator();
-	return 0;
-}
-
 namespace lua_race {
 
 	std::string register_metatable(lua_State * L)
@@ -120,14 +105,6 @@ namespace lua_race {
 
 		lua_pushstring(L, "race");
 		lua_setfield(L, -2, "__metatable");
-		
-		luaL_newmetatable(L, Gen);
-		static luaL_Reg const generator[] = {
-			{ "__call", &impl_name_generator_call},
-			{ "__gc", &impl_name_generator_collect},
-			{ nullptr, nullptr}
-		};
-		luaL_setfuncs(L, generator, 0);
 
 		return "Adding getrace metatable...\n";
 	}
